COMP: fix warning about int/double.
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=3724
git-svn-id: http://svn.slicer.org/Slicer4/trunk@3724 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/vtkMRMLDiffusionTensorDisplayPropertiesNode.h
@@ -337,7 +337,7 @@ class VTK_MRML_EXPORT vtkMRMLDiffusionTensorDisplayPropertiesNode : public vtkMR
   // Description:
   // Set the radius of the tube glyph
   //vtkSetMacro(TubeGlyphRadius, double);
-  void SetTubeGlyphRadius( int radius ) {
+  void SetTubeGlyphRadius( double radius ) {
 
     if ( this->TubeGlyphRadius != radius ) 
       {
